moved administrator to its own module
